file uploads go through fileUpload function
