+++ src/components/Recording/Recording.js
@@ -1,9 +1,10 @@
 import React from "react";
 import { ReactMic } from "react-mic";
-import axios from "axios";
+// import axios from "axios";
 import axiosAPI from "../../api/stoareeAPI";
 import { connect } from "react-redux";
 import { confirmUploadComplete, resetUploadStatus } from "../../redux/storyReducer";
+import fileUpload from "../../services/fileUpload";
 
 // CSS
 import "./../../css/main.css";
@@ -41,6 +42,16 @@ class Recording extends React.Component {
     console.log("chunk of real-time data is: ", recordedBlob);
   }
 
+  submitQuestion = (values, url) => {
+    // save question in story
+    axiosAPI.post(`/questions/${this.props.currentStory}`, {
+      ...values,
+      audioFileURL: url
+    }).then(res => {
+      this.props.confirmUploadComplete();
+    });
+  }
+
   onStop = (recordedBlob) => {
     console.log("recordedBlob is: ", recordedBlob);
 
@@ -49,43 +60,11 @@ class Recording extends React.Component {
 
     const fileName = Math.random().toString() + ".webm";
     const fileType = recordedBlob.options.mimeType;
-    console.log("Preparing the upload");
-    axiosAPI
-      .post("https://polar-castle-01694.herokuapp.com/sign_s3", {
-        fileName: fileName,
-        fileType: fileType
-      })
-      .then(response => {
-        const returnData = response.data.data.returnData;
-        const signedRequest = returnData.signedRequest;
-        console.log("Received a signed request " + signedRequest);
-        
-        // Put the fileType in the headers for the upload
-        const options = {
-          headers: {
-            "Content-Type": fileType
-          }
-        };
-        console.log(signedRequest);
-        console.log(options);
 
-        return axios
-          .put(signedRequest, recordedBlob.blob, options)
-          .then(result => {
-            console.log("Response from s3");
-
-            // save question in story
-            axiosAPI.post(`/questions/${this.props.currentStory}`, {
-              question: this.props.currentQuestion,
-              audioFileURL: returnData.url
-            }).then(response => {
-              this.props.confirmUploadComplete();
-            });
-          });
-      })
-      .catch(error => {
-        console.error(error);
-      });
+    const values = {
+      question: this.props.currentQuestion
+    }
+    fileUpload(values, recordedBlob.blob, this.submitQuestion, fileName, fileType);
   }
 
   render() {
