+++ src/pages/CreateStoryPage/CreateStoryPage.js
@@ -1,14 +1,14 @@
 import React from "react";
-import axios from "axios";
 import axiosAPI from "../../api/stoareeAPI";
 import { connect } from "react-redux";
 
 import { setCurrentStory } from "../../redux/storyReducer";
+import fileUpload from "../../services/fileUpload";
 
 import StoryForm from "../../components/StoryForm/StoryForm"
 import RecordPage from "../RecordPage/RecordPage";
 
-// CSS 
+// CSS
 import './CreateStoryPage.css';
 
 function mapStateToProps(state) {
@@ -38,32 +38,7 @@ class CreateStoryPage extends React.Component {
 
   onSubmit = (values) => {
     if (values.image) {
-      console.log('Preparing the upload');
-
-      axiosAPI.post("/sign_s3", {
-        fileName: Math.random().toString() + "/" + values.image.name,
-        fileType: values.image.type
-      }).then(response => {
-        const returnData = response.data.data.returnData;
-        const signedRequest = returnData.signedRequest;
-        const url = returnData.url;
-
-        console.log('Received a signed request ' + signedRequest);
-
-        const options = {
-          headers: {
-            'Content-Type': values.image.type
-          }
-        };
-
-        axios.put(signedRequest, values.image, options).then(result => {
-          console.log("Response from s3");
-
-          this.postStory(values, url);
-        });
-      }).catch(error => {
-        alert(JSON.stringify(error));
-      });
+      fileUpload(values, values.image, this.postStory);
     }
     else {
       this.postStory(values);
