+++ src/pages/EditStoryPage/EditStoryPage.js
@@ -1,7 +1,7 @@
 import React from "react";
-import axios from "axios";
 import axiosAPI from "../../api/stoareeAPI";
 import StoryForm from "../../components/StoryForm/StoryForm";
+import fileUpload from "../../services/fileUpload";
 
 class EditStoryPage extends React.Component {
   state = {
@@ -25,32 +25,7 @@ class EditStoryPage extends React.Component {
 
   onSubmit = (values) => {
     if (values.image) {
-      console.log('Preparing the upload');
-
-      axiosAPI.post("/sign_s3", {
-        fileName: Math.random().toString() + "/" + values.image.name,
-        fileType: values.image.type
-      }).then(response => {
-        const returnData = response.data.data.returnData;
-        const signedRequest = returnData.signedRequest;
-        const url = returnData.url;
-
-        console.log('Received a signed request ' + signedRequest);
-
-        const options = {
-          headers: {
-            'Content-Type': values.image.type
-          }
-        };
-
-        axios.put(signedRequest, values.image, options).then(result => {
-          console.log("Response from s3");
-
-          this.sendUpdate(values, url);
-        });
-      }).catch(error => {
-        alert(JSON.stringify(error));
-      });
+      fileUpload(values, values.image, this.sendUpdate);
     }
     else {
       this.sendUpdate(values);
